v0.1.7 (#20)
