+++ pkg/interface/package.json
@@ -1,6 +1,6 @@
 {
   "name": "interface",
-  "version": "0.1.6",
+  "version": "0.1.7",
   "description": "",
   "main": "index.js",
   "private": true,
